feat: allow users to pick local audio source location
context.getExternalFilesDir(null) will default to what it believes is the primary storage, sometimes that's internal and sometimes that's external. With this change users will now be able to pick which option they want. If the user does nothing it defaults back to the existing getExternalFilesDir and allows the system to pick the default.